+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaConnect.java
@@ -56,7 +56,12 @@ import static java.util.Collections.unmodifiableList;
                         )
                 },
                 subresources = @Crd.Spec.Subresources(
-                        status = @Crd.Spec.Subresources.Status()
+                        status = @Crd.Spec.Subresources.Status(),
+                        scale = @Crd.Spec.Subresources.Scale(
+                                specReplicasPath = KafkaConnect.SPEC_REPLICAS_PATH,
+                                statusReplicasPath = KafkaConnect.STATUS_REPLICAS_PATH,
+                                labelSelectorPath = KafkaConnect.LABEL_SELECTOR_PATH
+                        )
                 ),
                 additionalPrinterColumns = {
                         @Crd.Spec.AdditionalPrinterColumn(
@@ -93,6 +98,9 @@ public class KafkaConnect extends CustomResource implements UnknownPropertyPrese
     public static final String CRD_NAME = RESOURCE_PLURAL + "." + RESOURCE_GROUP;
     public static final String SHORT_NAME = "kc";
     public static final List<String> RESOURCE_SHORTNAMES = singletonList(SHORT_NAME);
+    public static final String SPEC_REPLICAS_PATH = ".spec.replicas";
+    public static final String STATUS_REPLICAS_PATH = ".status.replicas";
+    public static final String LABEL_SELECTOR_PATH = ".status.selector";
 
     private String apiVersion;
     private KafkaConnectSpec spec;