+++ api/src/main/java/io/strimzi/api/kafka/model/status/KafkaConnectStatus.java
@@ -6,9 +6,11 @@ package io.strimzi.api.kafka.model.status;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.strimzi.api.kafka.model.Constants;
 import io.strimzi.api.kafka.model.connect.ConnectorPlugin;
 import io.strimzi.crdgenerator.annotations.Description;
+import io.strimzi.crdgenerator.annotations.KubeLink;
 import io.sundr.builder.annotations.Buildable;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -31,6 +33,8 @@ public class KafkaConnectStatus extends Status {
 
     private String url;
     private List<ConnectorPlugin> connectorPlugins;
+    private int replicas;
+    private LabelSelector podSelector;
 
     @Description("The URL of the REST API endpoint for managing and monitoring Kafka Connect connectors.")
     public String getUrl() {
@@ -49,5 +53,26 @@ public class KafkaConnectStatus extends Status {
 
     public void setConnectorPlugins(List<ConnectorPlugin> connectorPlugins) {
         this.connectorPlugins = connectorPlugins;
+    }
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @Description("The current number of pods being used to provide this resource.")
+    public int getReplicas() {
+        return replicas;
+    }
+
+    public void setReplicas(int replicas) {
+        this.replicas = replicas;
+    }
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @KubeLink(group = "meta", version = "v1", kind = "labelselector")
+    @Description("Label selector for pods providing this resource.")
+    public LabelSelector getPodSelector() {
+        return podSelector;
+    }
+
+    public void setPodSelector(LabelSelector podSelector) {
+        this.podSelector = podSelector;
     }
 }