+++ api/src/test/java/io/strimzi/api/kafka/model/KafkaBridgeCrdIT.java
@@ -32,6 +32,11 @@ public class KafkaBridgeCrdIT extends AbstractCrdIT {
     }
 
     @Test
+    void testKafkaBridgeScaling() {
+        createScaleDelete(KafkaBridge.class, "KafkaBridge.yaml");
+    }
+
+    @Test
     void testKafkaBridgeMinimal() {
         createDelete(KafkaBridge.class, "KafkaBridge-minimal.yaml");
     }
@@ -109,7 +114,7 @@ public class KafkaBridgeCrdIT extends AbstractCrdIT {
     void setupEnvironment() {
         cluster.createNamespace(NAMESPACE);
         cluster.createCustomResources(TestUtils.CRD_KAFKA_BRIDGE);
-        cluster.cmdClient().waitForResourceCreation("crd", "kafkabridges.kafka.strimzi.io");
+        waitForCrd("crd", "kafkabridges.kafka.strimzi.io");
     }
 
     @AfterAll