+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaConnectAssemblyOperatorTest.java
@@ -191,6 +191,8 @@ public class KafkaConnectAssemblyOperatorTest {
                 KafkaConnectStatus connectStatus = capturedConnects.get(0).getStatus();
 
                 assertThat(connectStatus.getUrl(), is("http://foo-connect-api.test.svc:8083"));
+                assertThat(connectStatus.getReplicas(), is(connect.getReplicas()));
+                assertThat(connectStatus.getPodSelector().getMatchLabels(), is(connect.getSelectorLabels().toMap()));
                 assertThat(connectStatus.getConditions().get(0).getStatus(), is("True"));
                 assertThat(connectStatus.getConditions().get(0).getType(), is("Ready"));
 
@@ -851,6 +853,8 @@ public class KafkaConnectAssemblyOperatorTest {
                 assertThat(capturedConnects, hasSize(1));
                 KafkaConnectStatus connectStatus = capturedConnects.get(0).getStatus();
                 assertThat(connectStatus.getUrl(), is("http://foo-connect-api.test.svc:8083"));
+                assertThat(connectStatus.getReplicas(), is(connect.getReplicas()));
+                assertThat(connectStatus.getPodSelector().getMatchLabels(), is(connect.getSelectorLabels().toMap()));
                 assertThat(connectStatus.getConditions().get(0).getStatus(), is("True"));
                 assertThat(connectStatus.getConditions().get(0).getType(), is("Ready"));
                 if (connectorOperator) {